% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataNormalNs.R
\docType{data}
\name{dataNormalNs}
\alias{dataNormalNs}
\title{Sample Data for Small Area Estimation using Hierarchical Bayesian Method under Normal distribution with non-sampled areas}
\format{
A data frame with 50 rows and 5 variables:
\describe{
  \item{y}{Direct Estimation of y}
  \item{x1}{Auxiliary variable of x1}
  \item{x2}{Auxiliary variable of x2}
  \item{x3}{Auxiliary variable of x3}
  \item{vardir}{Sampling Variance of y}
}
}
\usage{
dataNormalNs
}
\description{
Dataset to simulate Small Area Estimation using Hierarchical Bayesian Method under Normal distribution with non-sampled areas

This data contains NA values that indicates no sampled at one or more small areas. It uses the dataNormal
with the direct estimates and the related variances in 5 small areas are missing.
}
\keyword{datasets}
