% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{Normal}
\alias{Normal}
\title{Small Area Estimation using Hierarchical Bayesian under Normal Distribution}
\usage{
Normal(
  formula,
  vardir,
  iter.update = 3,
  iter.mcmc = 2000,
  thin = 1,
  burn.in = 1000,
  data
)
}
\arguments{
\item{formula}{Formula that describe the fitted model}

\item{vardir}{Sampling variances of direct estimations}

\item{iter.update}{Number of updates with default \code{3}}

\item{iter.mcmc}{Number of total iterations per chain with default \code{2000}}

\item{thin}{Thinning rate, must be a positive integer with default \code{1}}

\item{burn.in}{Number of iterations to discard at the begining with default \code{1000}}

\item{data}{The data frame}
}
\value{

}
\description{
This function is implemented to variable of interest \eqn{(y)} that assumed to be a Normal Distribution. The range of data is \eqn{(-\infty < y < \infty)}
}
